Multiple bugfixes, including match against self
The nested method inside find_compare was pulling from same_name_files
rather than file_list. This means that name_index was compared twice
and size_index was not compared at all. To prevent this in the future,
added an exception for when a file is compared to itself and to check
the comparison cache for both (file_a, file_b) and (file_b, file_a).

Changed how indexes and files are printed. Also made indexes always
print inside main.

Fixed a spelling error with "unique" in comparison.